requirements.txt
Django==4.2.7
django-allauth==0.58.2
flower==2.0.1
humanize==4.8.0
celery==5.3.5
openai==1.2.4
redis==5.0.1
requests==2.31.0
sentry-sdk==1.35.0
stripe==7.6.0
